+++ packages/terra-menu/CHANGELOG.md
@@ -3,7 +3,11 @@ Changelog
 
 Unreleased
 ----------
-#Added
+
+### Fixed
+* Bug around selectable menus with only 1 child element.
+
+### Added
 * Added ability to disable menu items
 
 1.2.0 - (September 26, 2017)